Export and import from profile file

+++ import_profile.sh
@@ -1,28 +1,52 @@
 #!/bin/bash
 source var.sh
 
-#get the profile file
-#curl something
+#check if it's executed as root
+if [[ $EUID -ne 0 ]]; then
+  echo "You must be a root user" 2>&1
+  exit 1
+fi
 
-tar -xzvf $PROFILE_DIR.tar.gz
+PROFILE_NAME=profile
+while getopts ":f:" opt; do
+  case $opt in
+    f)
+      PROFILE_NAME=$OPTARG
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG" >&2
+      ;;
+  esac
+done
+
+tar -xzvf $PROFILE_NAME.tar.gz
+
 #copy the repositories of the profile
 cp -rf $PROFILE_DIR/$REPOS_DIR/** $YUM_REPO_PATH
 
 #install the packages specified in insalled_packages
-while read -r line
-do
-    name=$line
-    echo ""
-    echo "Install package "$name
-    eval yum install -y $name
-done < "$PROFILE_DIR/$PACKAGE_LIST_FILE"
+if [ -f $PROFILE_DIR/$PACKAGE_LIST_FILE ]
+then
+   while read -r line
+   do
+     name=$line
+     echo ""
+     echo "Install package "$name
+     eval yum install -y $name
+   done < "$PROFILE_DIR/$PACKAGE_LIST_FILE"
+fi
+
 
 #remove the packages mentioned in remove_packges
-filename=remove_packages
-while read -r line
-do
+if [ -f $PROFILE_DIR/$REMOVE_PACKAGE_LIST_FILE ]
+then
+  while read -r line
+  do
     name=$line
     echo ""
     echo "Remove package "$name
     eval yum remove -y $name
-done < "$PROFILE_DIR/$REMOVE_PACKAGE_LIST_FILE"
+  done < "$PROFILE_DIR/$REMOVE_PACKAGE_LIST_FILE"
+fi
+
+echo "Done"
